requirements.txt
selenium
beautifulsoup4
webdriver_manager
psycopg2-binary